Module PROTOCOL.Type_iv

An initialization vector

val construct : unit -> t

Create an initialization vector

val toBitstring : t -> t

val fromBitstring : t -> t

val xassert : t -> t

Verify that the object is a valid iv object.

raises Invalid_argument

when the object is not a valid iv object.