Module PROTOCOL.Type_recvoutput

The updates to your record of the other party that should be persisted after receiving a message

val construct : unit -> t record_recvoutput

Create a recvoutput object

val xassert : t record_recvoutput -> t record_recvoutput

Verify that the object is a valid recvoutput object.

raises Invalid_argument

when the object is not a valid recvoutput object.