Module PROTOCOL.Type_sendoutput

The updates to your record of the other party that should be persisted after sending a message

val construct : unit -> t record_sendoutput

Create a sendoutput object

val xassert : t record_sendoutput -> t record_sendoutput

Verify that the object is a valid sendoutput object.

raises Invalid_argument

when the object is not a valid sendoutput object.